+++ src/app/delivery/rest/search/search.service.ts
@@ -2,10 +2,10 @@ import { Inject, Injectable } from '@angular/core';
 import { SearchApi } from "./search-api";
 import { HttpClient } from "@angular/common/http";
 import { map, Observable } from "rxjs";
-import { SearchResult } from "../domain/search-result.model";
-import { SearchResultMapper } from "../data/search-result-mapper.service";
-import { GATEWAY_BASE_URL } from "../../config/gateway.config";
-import { MovieSearchResultListDto } from "../../delivery/rest/search/movie-search-result-list.dto";
+import { SearchResult } from "../../../shared/domain/search-result.model";
+import { SearchResultMapper } from "./search-result-mapper.service";
+import { GATEWAY_BASE_URL } from "../../../config/gateway.config";
+import { MovieSearchResultListDto } from "./movie-search-result-list.dto";
 
 @Injectable()
 export class SearchService implements SearchApi {
