+++ src/app/delivery/rest/user/register-user-api.ts
@@ -1,5 +1,5 @@
 import { Observable } from "rxjs";
-import { RegisterUserDto } from "../../delivery/rest/user/register-user.dto";
+import { RegisterUserDto } from "./register-user.dto";
 
 export interface RegisterUserApi {
   registerUser<T extends RegisterUserDto>(arg: T): Observable<any>;
