+++ src/app/delivery/rest/user/register-user.service.ts
@@ -3,9 +3,9 @@ import { AuthService, User } from "@auth0/auth0-angular";
 import { HttpClient } from "@angular/common/http";
 import { RegisterUserApi } from "./register-user-api";
 import {Observable, of} from 'rxjs';
-import {GATEWAY_BASE_URL} from "../../config/gateway.config";
-import {RegisterUserDto} from "../../delivery/rest/user/register-user.dto";
-import {UserDto} from "../../delivery/rest/user/user.dto";
+import {GATEWAY_BASE_URL} from "../../../config/gateway.config";
+import {RegisterUserDto} from "./register-user.dto";
+import {UserDto} from "./user.dto";
 
 // TODO:
 //  - this service is supposed to make a call to the gateway - which in turn will call the database/backend to
