+++ src/app/shared/pipes/truncate.pipe.spec.ts
