+++ src/app/shared/pipes/truncate.pipe.ts
