+++ src/app/shared/shared.module.ts
@@ -2,10 +2,10 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { SearchBarComponent } from './presentation/search-bar/search-bar.component';
 import { FormsModule } from "@angular/forms";
-import { SearchService } from "./delivery/search.service";
-import { SearchResultMapper } from "./data/search-result-mapper.service";
+import { SearchService } from "../delivery/rest/search/search.service";
+import { SearchResultMapper } from "../delivery/rest/search/search-result-mapper.service";
 import { SearchResultComponent } from './presentation/search-result/search-result.component';
-import { TruncatePipe } from './data/truncate.pipe';
+import { TruncatePipe } from './pipes/truncate.pipe';
 
 @NgModule({
   declarations: [
